requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
anthropic>=0.18.1
aiohttp>=3.8.1
pydantic>=1.8.2
mcp
